# sql-audit-service/sql-audit-service-web/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: SQL Audit Service API
  version: 2.0.0
  description: |
    REST API for querying SQL audit reports, managing checker configuration, and accessing statistics.

    The SQL Audit Service provides retrospective analysis of SQL execution patterns, risk scoring,
    and configuration management for audit checkers.

    **Key Features:**
    - Query audit findings with filtering and pagination
    - Access dashboard statistics and trends
    - Manage checker configuration
    - Export audit data in multiple formats

  contact:
    name: SQL Guard Team
    email: support@example.com
    url: https://github.com/example/sql-guard
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8090/api/v1
    description: Local development server
  - url: https://audit-dev.example.com/api/v1
    description: Development environment
  - url: https://audit.example.com/api/v1
    description: Production environment
...
paths:
...
  /configuration/checkers/{checkerId}:
    get:
      tags:
        - Configuration
      summary: Get checker configuration
      description: Retrieve configuration for a specific checker
      operationId: getCheckerConfig
      parameters:
        - name: checkerId
          in: path
          required: true
          schema:
            type: string
            example: "TABLE_LOCK"

      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckerConfig'

    put:
      tags:
        - Configuration
      summary: Update checker configuration
      description: Update configuration for a specific checker
      operationId: updateCheckerConfig
      parameters:
        - name: checkerId
          in: path
          required: true
          schema:
            type: string

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckerConfigUpdate'

      responses:
        '200':
          description: Configuration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckerConfig'
...
components:
  schemas:
...
    CheckerConfig:
      type: object
      description: Checker configuration details
      properties:
        checkerId:
          type: string
        enabled:
          type: boolean
        threshold:
          type: integer
        severityLevels:
          type: object
          additionalProperties: true
    CheckerConfigUpdate:
      type: object
      description: Checker configuration update request
      properties:
        enabled:
          type: boolean
        threshold:
          type: integer
        severityLevels:
          type: object
          additionalProperties: true
...